parser renamed to worker in order to avoid naming confilcts. Methos has been moved inside the object.
